+++ test/karcarthy/test_runner.clj
@@ -14,7 +14,8 @@
     karcarthy.runner-test
     karcarthy.orchestrate-test
     karcarthy.self-test
-    karcarthy.session-test])
+    karcarthy.session-test
+    karcarthy.otel-test])
 
 (defn -main [& _]
   (doseq [ns test-namespaces] (require ns))